requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
mixins==0.1.4
pkg-resources==0.0.0
python-dateutil==2.8.2
pytube==10.8.5
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
django-filter==2.2.0
drf-nested-routers==0.91.0